refactor: update privacy policy block types naming
