+++ app/types/store/pages/privacy-policy/blocks/introSectionBlock.ts
